+++ src/main/java/io/goobox/sync/sia/model/PriceInfo.java
@@ -23,43 +23,43 @@ import org.jetbrains.annotations.NotNull;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintWriter;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public class PriceInfo {
 
     @NotNull
-    private final BigDecimal download;
+    private final BigInteger download;
     @NotNull
-    private final BigDecimal upload;
+    private final BigInteger upload;
     @NotNull
-    private final BigDecimal storage;
+    private final BigInteger storage;
     @NotNull
-    private final BigDecimal contract;
+    private final BigInteger contract;
 
     public PriceInfo(@NotNull final InlineResponse20012 prices) {
-        this.download = new BigDecimal(prices.getDownloadterabyte());
-        this.upload = new BigDecimal(prices.getUploadterabyte());
-        this.storage = new BigDecimal(prices.getStorageterabytemonth());
-        this.contract = new BigDecimal(prices.getFormcontracts());
+        this.download = new BigInteger(prices.getDownloadterabyte());
+        this.upload = new BigInteger(prices.getUploadterabyte());
+        this.storage = new BigInteger(prices.getStorageterabytemonth());
+        this.contract = new BigInteger(prices.getFormcontracts());
     }
 
     @NotNull
-    public BigDecimal getDownload() {
+    public BigInteger getDownload() {
         return download;
     }
 
     @NotNull
-    public BigDecimal getUpload() {
+    public BigInteger getUpload() {
         return upload;
     }
 
     @NotNull
-    public BigDecimal getStorage() {
+    public BigInteger getStorage() {
         return storage;
     }
 
     @NotNull
-    public BigDecimal getContract() {
+    public BigInteger getContract() {
         return contract;
     }
